test(repo): parse each swept module once and follow an aliased base import (#15007)
The uncollected-class sweep parsed every module twice: once through the
guarded _parse_module and again with a bare ast.parse inside
uncollected_test_methods. The sweep now takes the tree it already has, and
_parse_source is the single place text becomes a tree, so no reachable call
path can parse unguarded.

A cross-module base was resolved by the name it is referenced by, then looked
up in the target module's dict, which is keyed by the real name -- so
'from mod import Foo as Bar' left Bar unresolvable. That failed safe but
produced a spurious unresolvable-base report. The alias is now followed, with
a fixture pinning both directions.

Splitting the model out of the guard is what the file-size ratchet requires:
the file was grandfathered at 921 lines and may not grow. It is now three
files, each under MAX_LINES, and the grandfathered entry is deleted from
KNOWN_LARGE and RATCHET_BASELINE.

Ratchet numbers unchanged: autobot-backend 77, autobot-frontend 10,
autobot-infrastructure 19, interface stubs 2, unresolvable bases 0.
