+++ repo_tests/python_file_size_ratchet_baseline.py
@@ -538,7 +538,6 @@ RATCHET_BASELINE: dict[str, int] = {
     "repo_tests/ci_dispatch_watchdog_test.py": 1317,
     "repo_tests/enum_union_guard_test.py": 808,
     "repo_tests/sys_modules_leak_guard.py": 1164,
-    "repo_tests/test_methods_in_uncollected_classes_test.py": 921,
     "scripts/audit_api_wiring.py": 897,
 }
